# contract/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
...
      "CursorDto" : {
        "description" : "Opaque pagination cursor. Read it from a response and send it back unchanged.",
        "type" : "string"
      },
...
      "PinIdsInputDto" : {
        "type" : "object",
        "required" : [ "pinIds" ],
        "properties" : {
          "pinIds" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid",
              "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
            },
            "maxItems" : 10000,
            "minItems" : 1
          }
        }
      },
...
    },
    "securitySchemes" : {
      "BearerScheme" : {
        "type" : "http",
        "description" : "Session token issued by POST /api/v1/sessions, sent as 'Authorization: Bearer <token>'.",
        "scheme" : "bearer"
      },
      "CookieScheme" : {
        "type" : "apiKey",
        "description" : "The same session token, issued by POST /api/v1/sessions with transport COOKIE and sent by the browser on its own.",
        "name" : "pinry_session",
        "in" : "cookie"
      }
    },
    "responses" : {
...
      "InvalidBatchBody" : {
        "description" : "The body is missing, or its list of identifiers is empty or longer than 10000",
        "content" : {
          "application/problem+json" : {
            "schema" : {
              "allOf" : [ {
                "$ref" : "#/components/schemas/ProblemDetail"
              } ],
              "properties" : {
                "code" : {
                  "enum" : [ "VALIDATION_ERROR", "MALFORMED_BODY" ]
                }
              }
            }
          }
        }
      },
...
      "PinForbidden" : {
        "description" : "A pin the request names belongs to another account",
        "content" : {
          "application/problem+json" : {
            "schema" : {
              "allOf" : [ {
                "$ref" : "#/components/schemas/ProblemDetail"
              } ],
              "properties" : {
                "code" : {
                  "enum" : [ "PIN_INSUFFICIENT_PERMISSIONS" ]
                }
              }
            }
          }
        }
      },
...
      "PinAlreadyRecycled" : {
        "description" : "The pin is in the recycle bin",
        "content" : {
          "application/problem+json" : {
            "schema" : {
              "allOf" : [ {
                "$ref" : "#/components/schemas/ProblemDetail"
              } ],
              "properties" : {
                "code" : {
                  "enum" : [ "PIN_ALREADY_SOFT_DELETED" ]
                }
              }
            }
          }
        }
      },
      "PinInBodyNotFound" : {
        "description" : "A pin the body names does not exist",
        "content" : {
          "application/problem+json" : {
            "schema" : {
              "allOf" : [ {
                "$ref" : "#/components/schemas/ProblemDetail"
              } ],
              "properties" : {
                "code" : {
                  "enum" : [ "PIN_DOES_NOT_EXIST" ]
                }
              }
            }
          }
        }
      },
...
    }
  },
  "paths" : {
...
    "/api/v1/pins" : {
      "delete" : {
        "summary" : "Recycle several pins, all or nothing",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PinIdsInputDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "Pins recycled"
          },
          "400" : {
            "$ref" : "#/components/responses/InvalidBatchBody"
          },
          "403" : {
            "$ref" : "#/components/responses/PinForbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/PinInBodyNotFound"
          },
          "409" : {
            "$ref" : "#/components/responses/PinAlreadyRecycled"
          },
          "415" : {
            "$ref" : "#/components/responses/UnsupportedMediaType"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthenticated"
          },
          "413" : {
            "$ref" : "#/components/responses/BodyTooLarge"
          }
        },
        "tags" : [ "Pin Controller" ],
        "security" : [ {
          "BearerScheme" : [ ]
        }, {
          "CookieScheme" : [ ]
        } ]
      },
      "get" : {
        "parameters" : [ {
          "name" : "cursor",
          "in" : "query",
          "schema" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/CursorDto"
            }, {
              "type" : "null"
            } ]
          }
        }, {
          "name" : "pageSize",
          "in" : "query",
          "schema" : {
            "type" : [ "integer", "null" ],
            "format" : "int32"
          }
        }, {
          "name" : "q",
          "in" : "query",
          "schema" : {
            "type" : [ "string", "null" ]
          }
        }, {
          "name" : "sort",
          "in" : "query",
          "schema" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/PinSortStrategyInputEnum"
            }, {
              "type" : "null"
            } ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PinListOutputDto"
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/BlankQuery"
          },
          "403" : {
            "$ref" : "#/components/responses/PinForbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/PinNotFound"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthenticated"
          }
        },
        "summary" : "List Pins",
        "tags" : [ "Pin Controller" ],
        "security" : [ {
          "BearerScheme" : [ ]
        }, {
          "CookieScheme" : [ ]
        } ]
      },
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PinCreationInputDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Pin created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PinOutputDto"
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/InvalidBody"
          },
          "415" : {
            "$ref" : "#/components/responses/UnsupportedMediaType"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthenticated"
          },
          "413" : {
            "$ref" : "#/components/responses/BodyTooLarge"
          }
        },
        "summary" : "Create Pin",
        "tags" : [ "Pin Controller" ],
        "security" : [ {
          "BearerScheme" : [ ]
        }, {
          "CookieScheme" : [ ]
        } ]
      }
    },
...
    "/api/v1/pins/{pinId}/media" : {
      "put" : {
        "summary" : "Set the pin's media (upload bytes, or request a server-side fetch)",
        "parameters" : [ {
          "name" : "pinId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "required" : [ "file" ],
                "properties" : {
                  "file" : {
                    "type" : "string",
                    "format" : "binary"
                  }
                }
              }
            },
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PinMediaDownloadInputDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Media created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MediaOutputDto"
                }
              }
            }
          },
          "200" : {
            "description" : "Media replaced",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MediaOutputDto"
                }
              }
            }
          },
          "400" : {
            "description" : "The upload has no file part, the body is not JSON or breaks a constraint, or the source URL is not an http or https address",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "allOf" : [ {
                    "$ref" : "#/components/schemas/ProblemDetail"
                  } ],
                  "properties" : {
                    "code" : {
                      "enum" : [ "VALIDATION_ERROR", "MALFORMED_BODY" ]
                    }
                  }
                }
              }
            }
          },
          "403" : {
            "$ref" : "#/components/responses/MediaForbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/MediaNotFound"
          },
          "413" : {
            "description" : "MEDIA_TOO_LARGE: the upload is past media.max_image_bytes for an image, media.max_video_bytes for a video. BODY_TOO_LARGE: the Content-Length is past quarkus.http.limits.max-body-size, which is above both; a chunked body past it gets a 413 with no body",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "allOf" : [ {
                    "$ref" : "#/components/schemas/ProblemDetail"
                  } ],
                  "properties" : {
                    "code" : {
                      "enum" : [ "MEDIA_TOO_LARGE", "BODY_TOO_LARGE" ]
                    }
                  }
                }
              }
            }
          },
          "415" : {
            "description" : "UNSUPPORTED_MEDIA_TYPE: the route does not read this Content-Type. MEDIA_CODEC_UNSUPPORTED: the upload is a format or a codec the server reads and does not store; for a video, the detail names the codec",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "allOf" : [ {
                    "$ref" : "#/components/schemas/ProblemDetail"
                  } ],
                  "properties" : {
                    "code" : {
                      "enum" : [ "UNSUPPORTED_MEDIA_TYPE", "MEDIA_CODEC_UNSUPPORTED" ]
                    }
                  }
                }
              }
            }
          },
          "422" : {
            "description" : "MEDIA_INVALID: the upload is neither an image nor a video the server reads, or it is past media.max_pixels_per_frame. MEDIA_TOO_LONG: the video lasts longer than media.max_video_seconds",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "allOf" : [ {
                    "$ref" : "#/components/schemas/ProblemDetail"
                  } ],
                  "properties" : {
                    "code" : {
                      "enum" : [ "MEDIA_INVALID", "MEDIA_TOO_LONG" ]
                    }
                  }
                }
              }
            }
          },
          "202" : {
            "description" : "Download accepted",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PinMediaStateDto"
                }
              }
            }
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthenticated"
          }
        },
        "tags" : [ "Media Controller" ],
        "security" : [ {
          "BearerScheme" : [ ]
        }, {
          "CookieScheme" : [ ]
        } ]
      },
      "get" : {
        "parameters" : [ {
          "name" : "pinId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        }, {
          "name" : "animated",
          "in" : "query",
          "schema" : {
            "type" : [ "boolean", "null" ]
          }
        }, {
          "name" : "size",
          "in" : "query",
          "schema" : {
            "type" : [ "string", "null" ]
          }
        }, {
          "name" : "If-None-Match",
          "in" : "header",
          "schema" : {
            "type" : [ "string", "null" ]
          },
          "required" : true
        }, {
          "name" : "Range",
          "in" : "header",
          "schema" : {
            "type" : [ "string", "null" ]
          },
          "required" : true
        } ],
        "responses" : {
          "200" : {
            "description" : "The original, or a WebP rendition",
            "content" : {
              "image/*" : { },
              "video/*" : { }
            }
          },
          "206" : {
            "description" : "The requested byte range of the original, Content-Range set",
            "content" : {
              "image/*" : { },
              "video/*" : { }
            }
          },
          "400" : {
            "description" : "The size names no rendition",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "allOf" : [ {
                    "$ref" : "#/components/schemas/ProblemDetail"
                  } ],
                  "properties" : {
                    "code" : {
                      "enum" : [ "MEDIA_RENDITION_SIZE_INVALID" ]
                    }
                  }
                }
              }
            }
          },
          "403" : {
            "$ref" : "#/components/responses/MediaForbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/MediaNotFound"
          },
          "416" : {
            "$ref" : "#/components/responses/RangeNotSatisfiable"
          },
          "422" : {
            "description" : "The media's frame is past media.max_pixels_per_frame, lowered since it was stored, or its decoder failed under media.decoder_timeout and media.decoder_memory_bytes in the last 24 hours",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "allOf" : [ {
                    "$ref" : "#/components/schemas/ProblemDetail"
                  } ],
                  "properties" : {
                    "code" : {
                      "enum" : [ "MEDIA_RENDITION_UNAVAILABLE" ]
                    }
                  }
                }
              }
            }
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthenticated"
          }
        },
        "summary" : "Get Media",
        "tags" : [ "Media Controller" ],
        "security" : [ {
          "BearerScheme" : [ ]
        }, {
          "CookieScheme" : [ ]
        } ]
      },
      "delete" : {
        "parameters" : [ {
          "name" : "pinId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "403" : {
            "$ref" : "#/components/responses/MediaForbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/MediaNotFound"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthenticated"
          }
        },
        "summary" : "Delete Media",
        "tags" : [ "Media Controller" ],
        "security" : [ {
          "BearerScheme" : [ ]
        }, {
          "CookieScheme" : [ ]
        } ]
      }
    },
    "/api/v1/pins/{pinId}/media/status" : {
      "get" : {
        "parameters" : [ {
          "name" : "pinId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PinMediaStateDto"
                }
              }
            }
          },
          "403" : {
            "$ref" : "#/components/responses/MediaForbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/MediaNotFound"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthenticated"
          }
        },
        "summary" : "Get Media Status",
        "tags" : [ "Media Controller" ],
        "security" : [ {
          "BearerScheme" : [ ]
        }, {
          "CookieScheme" : [ ]
        } ]
      }
    },
...
  },
  "info" : {
    "title" : "Pinry Reborn API",
    "version" : "25.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}